docs: migrate to storybook, closes #136 (#137)
