+++ README.adoc
@@ -17,17 +17,9 @@ toc::[]
 yarn add react-popover
 ```
 
+## Examples
 
-
-## Demos
-
-* link:https://littlebits.github.io/react-popover/build/playground.html[Playground]
-* link:https://littlebits.github.io/react-popover/build/rows.html[Rows]
-* link:https://littlebits.github.io/react-popover/build/jsx.html[Using JSX]
-
-To run demos locally execute `yarn start`. If you do not see logs in the console refresh the browser once (https://github.com/littlebits/react-popover/issues/35).
-
-
+Look at the link:https://littlebits.github.io/react-popover[stories in our storybook].
 
 ## API
 
@@ -98,4 +90,3 @@ Defines the size of the tip pointer.  Use .01 to disable tip.  Defaults to '7'.
 #### `appendTarget :: DOM selector (String)`
 
 - The DOM selector where this popover will be appended to. Default to 'body'.
-
