+++ lib/index.js
@@ -474,8 +474,4 @@ const Popover = createReactClass({
   },
 })
 
-
-
-// Support for CJS
-// http://stackoverflow.com/questions/33505992/babel-6-changes-how-it-exports-default
-module.exports = Popover
+export default Popover
