+++ package.json
@@ -3,14 +3,14 @@
   "version": "0.4.18",
   "scripts": {
     "clean": "rm -rf build",
-    "build-examples": "yarn run clean && webpack",
+    "deploy-storybook": "storybook-to-ghpages",
     "deploy-demos": "git checkout gh-pages && git merge master --no-edit && yarn run build-examples && git add -A && git commit -m 'Update demos' && git push && git checkout master",
     "build": "yarn run clean && babel --out-dir build lib",
-    "dev": "yarn start",
+    "dev": "start-storybook -p 6006",
     "test": "jest",
     "test-dev": "jest --watch",
-    "start": "webpack-dev-server --hot",
-    "postversion": "yarn run build && git push && git push --tags && npm publish && release"
+    "postversion": "yarn run build && git push && git push --tags && npm publish && release",
+    "build-storybook": "build-storybook"
   },
   "dependencies": {
     "create-react-class": "^15.5.3",
@@ -27,9 +27,13 @@
   "author": "Jason Kuhrt <jasonkuhrt@me.com> (http://jasonkuhrt.com/)",
   "name": "react-popover",
   "homepage": "https://github.com/littlebits/react-popover",
-  "main": "build/index.js",
+  "main": "index.js",
   "description": "A specification backed popover for react",
   "devDependencies": {
+    "@storybook/addon-actions": "^3.2.11",
+    "@storybook/addon-links": "^3.2.10",
+    "@storybook/react": "^3.2.11",
+    "@storybook/storybook-deployer": "^2.0.0",
     "babel-cli": "^6.3.15",
     "babel-core": "^6.3.15",
     "babel-eslint": "^7.1.1",
@@ -39,15 +43,13 @@
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-1": "^6.3.13",
     "bluebird": "^3.0.5",
-    "bootstrap-webpack-plugin": "^0.3.0",
     "chai": "^3.2.0",
-    "classnames": "^2.2.1",
-    "css-loader": "^0.17.0",
     "cssnext-loader": "^1.0.1",
     "eslint": "^3.17.1",
     "eslint-config-littlebits": "^0.6.2",
     "jest": "^19.0.2",
     "jest-cli": "^19.0.2",
+    "prettier": "^1.7.3",
     "ramda": "^0.18.0",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
@@ -55,10 +57,7 @@
     "react-hot-loader": "^1.2.4",
     "react-tappable": "^0.7.1",
     "react-test-renderer": "^15.4.2",
-    "release": "^1.1.7",
-    "style-loader": "^0.12.3",
-    "webpack": "^1.12.1",
-    "webpack-dev-server": "^1.10.1"
+    "release": "^1.1.7"
   },
   "bugs": {
     "url": "https://github.com/littlebits/react-popover/issues"
