+++ stories/rows/main.css
@@ -1,6 +1,3 @@
-@import ../base;
-
-
 
 /* Boiler Plate */
